# perf: implement size hints for `Rect` iterators