--- a/src/layout/rect/iter.rs
+++ b/src/layout/rect/iter.rs
@@ -37,16 +37,14 @@ impl Iterator for Rows {
     }
 
     fn size_hint(&self) -> (usize, Option<usize>) {
-        let remaining_count = self
+        let count = self
             .rect
             .height
-            // Subtract the number of rows that were provided from the start of the `Rect`, from the
-            // total height.
             .saturating_sub(self.current_row_fwd.saturating_sub(self.rect.y))
-            // Subtract the number of rows that were provided from the end of the `Rect`, from the
-            // total height.
-            .saturating_sub(self.rect.bottom().saturating_sub(self.current_row_back));
-        (remaining_count as usize, None)
+            .saturating_sub(self.rect.bottom().saturating_sub(self.current_row_back))
+            as usize;
+
+        (count, Some(count))
     }
 }
 
@@ -101,14 +99,13 @@ impl Iterator for Columns {
     }
 
     fn size_hint(&self) -> (usize, Option<usize>) {
-        let remaining = self
+        let count = self
             .rect
             .width
-            // Subtract the number of columns provided from the start, from the total width.
             .saturating_sub(self.current_column_fwd.saturating_sub(self.rect.x))
-            // Subtract the number of columns provided from the end, from the total width.
-            .saturating_sub(self.rect.right().saturating_sub(self.current_column_back));
-        (remaining as usize, None)
+            .saturating_sub(self.rect.right().saturating_sub(self.current_column_back))
+            as usize;
+        (count, Some(count))
     }
 }
 
@@ -166,7 +163,7 @@ impl Iterator for Positions {
     }
 
     fn size_hint(&self) -> (usize, Option<usize>) {
-        // Number of the remaining rows including the current row.
+        // Includes the current row.
         let remaining_rows = self
             .rect
             .height
@@ -179,17 +176,15 @@ impl Iterator for Positions {
                 .saturating_add(self.rect.x)
                 .saturating_sub(self.current_position.x)
         } else {
-            return (0, None);
+            return (0, Some(0));
         };
         // Decrement the remaining rows by one since we do not want to include the
         // current row.
         let remaining_rows_cell_count = remaining_rows
             .saturating_sub(1)
             .saturating_mul(self.rect.width);
-        (
-            remaining_cells.saturating_add(remaining_rows_cell_count) as usize,
-            None,
-        )
+        let count = remaining_cells.saturating_add(remaining_rows_cell_count) as usize;
+        (count, Some(count))
     }
 }
 
@@ -201,106 +196,106 @@ mod tests {
     fn rows() {
         let rect = Rect::new(0, 0, 2, 3);
         let mut rows = Rows::new(rect);
-        assert_eq!(rows.size_hint().0, 3);
+        assert_eq!(rows.size_hint(), (3, Some(3)));
         assert_eq!(rows.next(), Some(Rect::new(0, 0, 2, 1)));
-        assert_eq!(rows.size_hint().0, 2);
+        assert_eq!(rows.size_hint(), (2, Some(2)));
         assert_eq!(rows.next(), Some(Rect::new(0, 1, 2, 1)));
-        assert_eq!(rows.size_hint().0, 1);
+        assert_eq!(rows.size_hint(), (1, Some(1)));
         assert_eq!(rows.next(), Some(Rect::new(0, 2, 2, 1)));
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
         assert_eq!(rows.next(), None);
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
         assert_eq!(rows.next_back(), None);
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
     }
 
     #[test]
     fn rows_back() {
         let rect = Rect::new(0, 0, 2, 3);
         let mut rows = Rows::new(rect);
-        assert_eq!(rows.size_hint().0, 3);
+        assert_eq!(rows.size_hint(), (3, Some(3)));
         assert_eq!(rows.next_back(), Some(Rect::new(0, 2, 2, 1)));
-        assert_eq!(rows.size_hint().0, 2);
+        assert_eq!(rows.size_hint(), (2, Some(2)));
         assert_eq!(rows.next_back(), Some(Rect::new(0, 1, 2, 1)));
-        assert_eq!(rows.size_hint().0, 1);
+        assert_eq!(rows.size_hint(), (1, Some(1)));
         assert_eq!(rows.next_back(), Some(Rect::new(0, 0, 2, 1)));
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
         assert_eq!(rows.next_back(), None);
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
         assert_eq!(rows.next(), None);
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
     }
 
     #[test]
     fn rows_meet_in_the_middle() {
         let rect = Rect::new(0, 0, 2, 4);
         let mut rows = Rows::new(rect);
-        assert_eq!(rows.size_hint().0, 4);
+        assert_eq!(rows.size_hint(), (4, Some(4)));
         assert_eq!(rows.next(), Some(Rect::new(0, 0, 2, 1)));
-        assert_eq!(rows.size_hint().0, 3);
+        assert_eq!(rows.size_hint(), (3, Some(3)));
         assert_eq!(rows.next_back(), Some(Rect::new(0, 3, 2, 1)));
-        assert_eq!(rows.size_hint().0, 2);
+        assert_eq!(rows.size_hint(), (2, Some(2)));
         assert_eq!(rows.next(), Some(Rect::new(0, 1, 2, 1)));
-        assert_eq!(rows.size_hint().0, 1);
+        assert_eq!(rows.size_hint(), (1, Some(1)));
         assert_eq!(rows.next_back(), Some(Rect::new(0, 2, 2, 1)));
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
         assert_eq!(rows.next(), None);
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
         assert_eq!(rows.next_back(), None);
-        assert_eq!(rows.size_hint().0, 0);
+        assert_eq!(rows.size_hint(), (0, Some(0)));
     }
 
     #[test]
     fn columns() {
         let rect = Rect::new(0, 0, 3, 2);
         let mut columns = Columns::new(rect);
-        assert_eq!(columns.size_hint().0, 3);
+        assert_eq!(columns.size_hint(), (3, Some(3)));
         assert_eq!(columns.next(), Some(Rect::new(0, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 2);
+        assert_eq!(columns.size_hint(), (2, Some(2)));
         assert_eq!(columns.next(), Some(Rect::new(1, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 1);
+        assert_eq!(columns.size_hint(), (1, Some(1)));
         assert_eq!(columns.next(), Some(Rect::new(2, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
         assert_eq!(columns.next(), None);
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
         assert_eq!(columns.next_back(), None);
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
     }
 
     #[test]
     fn columns_back() {
         let rect = Rect::new(0, 0, 3, 2);
         let mut columns = Columns::new(rect);
-        assert_eq!(columns.size_hint().0, 3);
+        assert_eq!(columns.size_hint(), (3, Some(3)));
         assert_eq!(columns.next_back(), Some(Rect::new(2, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 2);
+        assert_eq!(columns.size_hint(), (2, Some(2)));
         assert_eq!(columns.next_back(), Some(Rect::new(1, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 1);
+        assert_eq!(columns.size_hint(), (1, Some(1)));
         assert_eq!(columns.next_back(), Some(Rect::new(0, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
         assert_eq!(columns.next_back(), None);
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
         assert_eq!(columns.next(), None);
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
     }
 
     #[test]
     fn columns_meet_in_the_middle() {
         let rect = Rect::new(0, 0, 4, 2);
         let mut columns = Columns::new(rect);
-        assert_eq!(columns.size_hint().0, 4);
+        assert_eq!(columns.size_hint(), (4, Some(4)));
         assert_eq!(columns.next(), Some(Rect::new(0, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 3);
+        assert_eq!(columns.size_hint(), (3, Some(3)));
         assert_eq!(columns.next_back(), Some(Rect::new(3, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 2);
+        assert_eq!(columns.size_hint(), (2, Some(2)));
         assert_eq!(columns.next(), Some(Rect::new(1, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 1);
+        assert_eq!(columns.size_hint(), (1, Some(1)));
         assert_eq!(columns.next_back(), Some(Rect::new(2, 0, 1, 2)));
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
         assert_eq!(columns.next(), None);
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
         assert_eq!(columns.next_back(), None);
-        assert_eq!(columns.size_hint().0, 0);
+        assert_eq!(columns.size_hint(), (0, Some(0)));
     }
 
     /// We allow a total of `65536` columns in the range `(0..=65535)`.  In this test we iterate