Implement `Iter::size_hint` for `Rows`, `Columns`, and `Positions`.

The `size_hint` also accounts for the `DoubleEndedIterator` implementation, and provides an accurate hint when using `Iter`, and `DoubleEndedIterator`.

Performance wise this doesn't change iteration in any way. However it optimizes collecting the items in a collection (`Vec`, `HashSet`, `HashMap`). 

<details>

<summary>Performance can be measured  by applying following patch, and then running "cargo bench rect":</summary>

```patch
diff --git a/benches/main/rect.rs b/benches/main/rect.rs
index 828bfc3..f273454 100644
--- a/benches/main/rect.rs
+++ b/benches/main/rect.rs
@@ -1,24 +1,113 @@
-use criterion::{black_box, criterion_group, BenchmarkId, Criterion};
+use criterion::{black_box, criterion_group, BatchSize, Bencher, BenchmarkId, Criterion};
 use ratatui::layout::Rect;

-fn rect_rows_benchmark(c: &mut Criterion) {
-    let rect_sizes = vec![
-        Rect::new(0, 0, 1, 16),
-        Rect::new(0, 0, 1, 1024),
-        Rect::new(0, 0, 1, 65535),
-    ];
-    let mut group = c.benchmark_group("rect_rows");
-    for rect in rect_sizes {
-        group.bench_with_input(BenchmarkId::new("rows", rect.height), &rect, |b, rect| {
-            b.iter(|| {
-                for row in rect.rows() {
-                    // Perform any necessary operations on each row
-                    black_box(row);
-                }
-            });
-        });
+fn rect_iters_benchmark(c: &mut Criterion) {
+    let rect_sizes = vec![[16, 16], [64, 64], [255, 255]];
+    let mut group = c.benchmark_group("rect");
+    // We don't use `Rect::new` as it will clamp the value to a max of 255x255.
+    for rect in rect_sizes.into_iter().map(|[width, height]| Rect {
+        width,
+        height,
+        ..Default::default()
+    }) {
+        group.bench_with_input(
+            BenchmarkId::new("rect_rows_iter", rect.height),
+            &rect,
+            |b, rect| rect_rows_iter(b, *rect),
+        );
+        group.bench_with_input(
+            BenchmarkId::new("rect_rows_collect", rect.height),
+            &rect,
+            |b, rect| rect_rows_collect(b, *rect),
+        );
+        group.bench_with_input(
+            BenchmarkId::new("rect_columns_iter", rect.width),
+            &rect,
+            |b, rect| rect_columns_iter(b, *rect),
+        );
+        group.bench_with_input(
+            BenchmarkId::new("rect_columns_collect", rect.width),
+            &rect,
+            |b, rect| rect_columns_collect(b, *rect),
+        );
+        group.bench_with_input(
+            BenchmarkId::new(
+                "rect_positions_iter",
+                format!("{}x{}", rect.width, rect.height),
+            ),
+            &rect,
+            |b, rect| rect_positions_iter(b, *rect),
+        );
+        group.bench_with_input(
+            BenchmarkId::new(
+                "rect_positions_collect",
+                format!("{}x{}", rect.width, rect.height),
+            ),
+            &rect,
+            |b, rect| rect_positions_collect(b, *rect),
+        );
     }
     group.finish();
 }

-criterion_group!(benches, rect_rows_benchmark);
+fn rect_rows_iter(c: &mut Bencher, rect: Rect) {
+    c.iter_batched(
+        || black_box(rect),
+        |rect| {
+            for row in black_box(rect.rows()) {
+                black_box(row);
+            }
+        },
+        BatchSize::LargeInput,
+    );
+}
+
+fn rect_rows_collect(c: &mut Bencher, rect: Rect) {
+    c.iter_batched(
+        || black_box(rect),
+        |rect| black_box(rect.rows()).collect::<Vec<_>>(),
+        BatchSize::LargeInput,
+    );
+}
+
+fn rect_columns_iter(c: &mut Bencher, rect: Rect) {
+    c.iter_batched(
+        || black_box(rect),
+        |rect| {
+            for col in black_box(rect.columns()) {
+                black_box(col);
+            }
+        },
+        BatchSize::LargeInput,
+    );
+}
+
+fn rect_columns_collect(c: &mut Bencher, rect: Rect) {
+    c.iter_batched(
+        || black_box(rect),
+        |rect| black_box(rect.columns()).collect::<Vec<_>>(),
+        BatchSize::LargeInput,
+    );
+}
+
+fn rect_positions_iter(c: &mut Bencher, rect: Rect) {
+    c.iter_batched(
+        || black_box(rect),
+        |rect| {
+            for pos in black_box(rect.positions()) {
+                black_box(pos);
+            }
+        },
+        BatchSize::LargeInput,
+    );
+}
+
+fn rect_positions_collect(b: &mut Bencher, rect: Rect) {
+    b.iter_batched(
+        || black_box(rect),
+        |rect| black_box(rect.positions()).collect::<Vec<_>>(),
+        BatchSize::LargeInput,
+    );
+}
+
+criterion_group!(benches, rect_iters_benchmark);
```
</details>

The exact performance in actual usage can vary. This change improves the performance of collecting the items by around 50-70%. With a u16::MAX provided to any of the iterators this drops substantially. This only drastically improved performance for lower inputs such as 1024 items. 

<details>

<summary>Benchmark result compared to without size hint</summary>

```
rect/rect_rows_iter/16  time:   [5.3884 ns 5.3892 ns 5.3904 ns]
                        change: [+0.0051% +0.1901% +0.3621%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 20 outliers among 100 measurements (20.00%)
  3 (3.00%) high mild
  17 (17.00%) high severe
rect/rect_rows_collect/16
                        time:   [24.130 ns 24.162 ns 24.191 ns]
                        change: [-49.206% -48.831% -48.432%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  6 (6.00%) high mild
  9 (9.00%) high severe
rect/rect_columns_iter/16
                        time:   [6.8879 ns 6.8956 ns 6.9038 ns]
                        change: [-1.2312% -1.0829% -0.9368%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
rect/rect_columns_collect/16
                        time:   [25.824 ns 25.849 ns 25.879 ns]
                        change: [-50.596% -50.254% -49.914%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe
Benchmarking rect/rect_positions_iter/16x16: Collecting 100 samples in estimated 5.0003 s (29M iterationsrect/rect_positions_iter/16x16
                        time:   [172.08 ns 172.09 ns 172.10 ns]
                        change: [-0.2681% -0.1776% -0.0955%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe
Benchmarking rect/rect_positions_collect/16x16: Collecting 100 samples in estimated 5.0009 s (24M iteratirect/rect_positions_collect/16x16
                        time:   [200.75 ns 200.95 ns 201.23 ns]
                        change: [-46.206% -45.987% -45.763%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) low severe
  1 (1.00%) low mild
  3 (3.00%) high severe
rect/rect_rows_iter/64  time:   [14.952 ns 14.959 ns 14.964 ns]
                        change: [-0.5365% -0.1234% +0.1449%] (p = 0.61 > 0.05)
                        No change in performance detected.
rect/rect_rows_collect/64
                        time:   [55.562 ns 55.615 ns 55.676 ns]
                        change: [-58.737% -58.527% -58.228%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe
rect/rect_columns_iter/64
                        time:   [17.975 ns 17.993 ns 18.008 ns]
                        change: [-1.0430% -0.7802% -0.4111%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
rect/rect_columns_collect/64
                        time:   [60.467 ns 60.506 ns 60.554 ns]
                        change: [-61.007% -60.821% -60.621%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe
Benchmarking rect/rect_positions_iter/64x64: Collecting 100 samples in estimated 5.0084 s (1.9M iterationrect/rect_positions_iter/64x64
                        time:   [2.6433 µs 2.6439 µs 2.6447 µs]
                        change: [-0.7008% -0.2778% -0.0235%] (p = 0.10 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe
Benchmarking rect/rect_positions_collect/64x64: Collecting 100 samples in estimated 5.0062 s (1.9M iteratrect/rect_positions_collect/64x64
                        time:   [2.6869 µs 2.6874 µs 2.6879 µs]
                        change: [-15.808% -15.498% -15.270%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  8 (8.00%) low severe
  5 (5.00%) low mild
  4 (4.00%) high mild
  2 (2.00%) high severe
rect/rect_rows_iter/255 time:   [55.839 ns 55.857 ns 55.879 ns]
                        change: [-0.0632% +0.0407% +0.1614%] (p = 0.54 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  6 (6.00%) high mild
  6 (6.00%) high severe
rect/rect_rows_collect/255
                        time:   [173.84 ns 173.99 ns 174.18 ns]
                        change: [-40.416% -40.260% -40.113%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
rect/rect_columns_iter/255
                        time:   [58.862 ns 58.895 ns 58.949 ns]
                        change: [+0.8115% +0.9726% +1.1506%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) high mild
  8 (8.00%) high severe
rect/rect_columns_collect/255
                        time:   [199.95 ns 200.09 ns 200.25 ns]
                        change: [-40.872% -40.593% -40.327%] (p = 0.00 < 0.05)
                        Performance has improved.
Benchmarking rect/rect_positions_iter/255x255: Collecting 100 samples in estimated 5.0715 s (121k iteratirect/rect_positions_iter/255x255
                        time:   [41.845 µs 41.846 µs 41.847 µs]
                        change: [-0.4778% -0.4015% -0.3222%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe
Benchmarking rect/rect_positions_collect/255x255: Collecting 100 samples in estimated 5.1011 s (121k iterrect/rect_positions_collect/255x255
                        time:   [42.034 µs 42.046 µs 42.061 µs]
                        change: [-7.5948% -7.4331% -7.2038%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe
```
</details>