@joshka 2024-10-15
Feel free to add benchmarks, but as a general rule, only worry about using multiple sizes if the absolute magnitude of the different sizes is on a meaningful human relevant scale (e.g. some amount of milliseconds would be relevant, some amount of nanoseconds would not). For rect related stuff, I'd choose maybe two values 200x50 (a neat terminal size) and u16::Max (this is likely not relevant to normal use cases, but it covers the worst case).

The easiest approach to benchmark comparisons is to put the bench in a single commit stacked before the changes (use interactive rebase is needed to sort this out / use fixup commits and autosquash to update). Then run the bench before then checkout the last commit and run again. Criterion gives nice deltas.