@airblast-dev 2024-10-15
I specifically removed the `u16::MAX` from the benchmark as part of the patch. This was due to `Rect::new` clamping, and `Rect::positions` taking pretty long on my machine when provided a `u16::MAX`. We can lower the sample count, or use a lower value than `u16::MAX`, but I am more inclined to use realistic values instead.

As for the benchmark and results, I will update my initial comment in a bit.