@airblast-dev 2024-10-15
> ...42µs with non-representative data...
I definitely agree that the performance isn't anything to write home about. Though for relatively simple iterators I think size_hint is a reasonable thing to implement.

Will add the benchmarks shortly.